core/training/train_safe_rl.py
"""
Safe RL baseline with constraint satisfaction for comparison.
"""

import torch
import torch.nn as nn
import numpy as np
import yaml
import argparse
from pathlib import Path
from tqdm import tqdm

from core.envs.urban_grid_env import UrbanGridEnv


class SafetyConstraint:
    """Safety constraints for drone operations."""
    
    def __init__(self, min_distance: float = 1.0, max_velocity: float = 2.0):
        self.min_distance = min_distance
        self.max_velocity = max_velocity
    
    def check_collision_constraint(self, positions: np.ndarray) -> bool:
        """Check if inter-drone distances satisfy safety constraint."""
        num_drones = positions.shape[0]
        for i in range(num_drones):
            for j in range(i + 1, num_drones):
                dist = np.linalg.norm(positions[i] - positions[j])
                if dist < self.min_distance:
                    return False
        return True
    
    def check_velocity_constraint(self, velocities: np.ndarray) -> bool:
        """Check if velocities are within safe limits."""
        return np.all(np.linalg.norm(velocities, axis=1) <= self.max_velocity)
    
    def compute_constraint_violation(self, positions: np.ndarray, velocities: np.ndarray) -> float:
        """Compute degree of constraint violation."""
        violation = 0.0
        
        # Collision constraint
        num_drones = positions.shape[0]
        for i in range(num_drones):
            for j in range(i + 1, num_drones):
                dist = np.linalg.norm(positions[i] - positions[j])
                if dist < self.min_distance:
                    violation += (self.min_distance - dist)
        
        # Velocity constraint
        for vel in velocities:
            speed = np.linalg.norm(vel)
            if speed > self.max_velocity:
                violation += (speed - self.max_velocity)
        
        return violation


def train_safe_rl(config_path: str):
    """
    Train with safety constraints using Lagrangian relaxation.
    
    This is a placeholder for more advanced safe RL methods like:
    - Constrained Policy Optimization (CPO)
    - Safety Layer approaches
    - Model Predictive Control (MPC) with constraints
    """
    print("Safe RL training - Placeholder implementation")
    print("For production, implement CPO or similar constrained optimization methods")
    
    # Load config
    with open(config_path, 'r') as f:
        config = yaml.safe_load(f)
    
    # Create environment
    env = UrbanGridEnv()
    safety_constraint = SafetyConstraint()
    
    # Training loop would go here with safety constraints
    # This is left as a template for future implementation
    
    print("Safe RL baseline complete")


if __name__ == '__main__':
    parser = argparse.ArgumentParser()
    parser.add_argument('--config', type=str, default='configs/base.yaml')
    args = parser.parse_args()
    
    train_safe_rl(args.config)
