requirements.txt
# Fed-MARL: Federated Multi-Agent Reinforcement Learning for Drone Swarms
# Core dependencies for training and evaluation

# =============================================================================
# Core ML/AI Framework
# =============================================================================
torch>=1.9.0
torchvision>=0.10.0
torchaudio>=0.9.0
numpy>=1.21.0
scipy>=1.7.0

# =============================================================================
# Reinforcement Learning
# =============================================================================
gymnasium>=0.28.0
gym>=0.21.0
stable-baselines3>=1.6.0

# =============================================================================
# Visualization & Plotting
# =============================================================================
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
pillow>=8.3.0

# =============================================================================
# Configuration & Data
# =============================================================================
pyyaml>=6.0
h5py>=3.1.0
pandas>=1.3.0
tqdm>=4.62.0

# =============================================================================
# Scientific Computing
# =============================================================================
scikit-learn>=1.0.0
scikit-image>=0.18.0

# =============================================================================
# Utilities
# =============================================================================
click>=8.0.0
colorama>=0.4.4
rich>=12.0.0

# =============================================================================
# Optional: Performance & Monitoring
# =============================================================================
# tensorboard>=2.8.0
# wandb>=0.12.0
# mlflow>=1.20.0

# =============================================================================
# Optional: Federated Learning (Future)
# =============================================================================
# flwr>=1.3.0
# syft>=0.6.0

# =============================================================================
# Development Dependencies (Optional)
# =============================================================================
# pytest>=6.2.0
# pytest-cov>=3.0.0
# black>=21.0.0
# isort>=5.9.0
# flake8>=4.0.0
# mypy>=0.910